Rename hooks/token -> hooks/balance
